cont.

+++ mojave/code/modules/locks/obj_defines.dm
@@ -8,6 +8,9 @@
 	var/can_be_picked = TRUE
 	//check if obj has lock and can be picked
 	var/lock_locked = FALSE //essentially - "has lock"
+	//for locks built into the object, that cant be taken off, pre-spawn etc.
+	var/built_in_lock = FALSE
+	var/built_in_lock_type = /obj/item/ms13/lock // replace with custom faction locks etc. which will be randomised on init
 	//moves the lock to the obj so it can be taken off etc.
 	var/obj/item/ms13/lock/lock
 
@@ -16,6 +19,20 @@
 	for(var/datum/component/storage/storage as anything in src.GetComponents(/datum/component/storage))
 		storage.locked = FALSE //unlocks the storage
 	return TRUE
+
+/obj/proc/update_lock(pre_spawn)
+	if(pre_spawn)
+		built_in_lock = TRUE
+		lock = new built_in_lock_type(src)
+		can_be_picked = TRUE
+		lock.item_lock_locked = TRUE
+		lock.lock_open = FALSE
+		lock_locked = TRUE
+		lock.moveToNullspace()
+		for(var/datum/component/storage/storage as anything in GetComponents(/datum/component/storage))
+			storage.locked = TRUE //locks
+		update_appearance()
+		return
 
 //for denying access/interaction with general locked items
 /obj/attack_hand(mob/user, list/modifiers)
@@ -35,47 +52,67 @@
 //for item interaction overrides on all general objects for placing locked items
 /obj/attackby(obj/item/I, mob/living/user, params) //good lordy this is some bad code, didnt know much back in 2022 and I can confirm this is levels of slopitude that I didnt know possible from me, but hell it works
 	//key interactions
-	if(I.item_flags & KEY_ITEM && lock)
-		var/obj/item/ms13/key/key = I
-		//alignment checks
-		if(key.bitt_1 != lock.pin_1)
-			playsound(src, 'mojave/sound/ms13effects/lockpicking/lockpick_tension_11.ogg', 50, TRUE)
-			to_chat(user, span_notice("The key dosen't turn in [name]."))
-			return
-		if(key.bitt_2 != lock.pin_2)
-			playsound(src, 'mojave/sound/ms13effects/lockpicking/lockpick_tension_11.ogg', 50, TRUE)
-			to_chat(user, span_notice("The key dosen't turn in [name]."))
-			return
-		if(key.bitt_3 != lock.pin_3)
-			playsound(src, 'mojave/sound/ms13effects/lockpicking/lockpick_tension_11.ogg', 50, TRUE)
-			to_chat(user, span_notice("The key dosen't turn in [name]."))
-			return
-		if(key.bitt_4 != lock.pin_4)
-			playsound(src, 'mojave/sound/ms13effects/lockpicking/lockpick_tension_11.ogg', 50, TRUE)
-			to_chat(user, span_notice("The key dosen't turn in [name]."))
-			return
-		if(key.bitt_5 != lock.pin_5)
-			playsound(src, 'mojave/sound/ms13effects/lockpicking/lockpick_tension_11.ogg', 50, TRUE)
-			to_chat(user, span_notice("The key dosen't turn in [name]."))
-			return
-		if(key.bitt_6 != lock.pin_6)
-			playsound(src, 'mojave/sound/ms13effects/lockpicking/lockpick_tension_11.ogg', 50, TRUE)
-			to_chat(user, span_notice("The key dosen't turn in [name]."))
-			return
-		if(!lock.item_lock_locked)
-			playsound(src, 'mojave/sound/ms13effects/key_lock.ogg', 50, TRUE)
-			to_chat(user, span_notice("You lock the [name]."))
-			lock.item_lock_locked = TRUE
-			return
-		if(lock.item_lock_locked)
-			playsound(src, 'mojave/sound/ms13effects/key_unlock.ogg', 50, TRUE)
-			to_chat(user, span_notice("You unlock the [name]."))
-			lock.item_lock_locked = FALSE
+	if(I.ms13_flags_1 & KEY_ITEM || LOCKING_ITEM && ms13_flags_1 & LOCKABLE_1) //lil more soul for yah
+		if(I.ms13_flags_1 & LOCKING_ITEM && lock)
+			if(I == src)
+				return //don't do an animation if attacking self
+			var/pixel_x_diff = 0
+			var/pixel_y_diff = 0
+
+			var/direction = get_dir(user, src)
+			if(direction & NORTH)
+				pixel_y_diff = 8
+			else if(direction & SOUTH)
+				pixel_y_diff = -8
+
+			if(direction & EAST)
+				pixel_x_diff = 8
+			else if(direction & WEST)
+				pixel_x_diff = -8
+			animate(user, pixel_x = user.pixel_x + pixel_x_diff, pixel_y = user.pixel_y + pixel_y_diff, time = 1, easing=BACK_EASING|EASE_IN, flags = ANIMATION_PARALLEL)
+			animate(user, pixel_x = user.pixel_x - pixel_x_diff, pixel_y = user.pixel_y - pixel_y_diff, time = 2, easing=SINE_EASING, flags = ANIMATION_PARALLEL)
+		if(I.ms13_flags_1 & KEY_ITEM && lock)
+			var/obj/item/ms13/key/key = I
+			//alignment checks
+			if(key.bitt_1 != lock.pin_1)
+				playsound(src, 'mojave/sound/ms13effects/lockpicking/lockpick_tension_11.ogg', 50, TRUE)
+				to_chat(user, span_notice("The key dosen't turn in [name]."))
+				return
+			if(key.bitt_2 != lock.pin_2)
+				playsound(src, 'mojave/sound/ms13effects/lockpicking/lockpick_tension_11.ogg', 50, TRUE)
+				to_chat(user, span_notice("The key dosen't turn in [name]."))
+				return
+			if(key.bitt_3 != lock.pin_3)
+				playsound(src, 'mojave/sound/ms13effects/lockpicking/lockpick_tension_11.ogg', 50, TRUE)
+				to_chat(user, span_notice("The key dosen't turn in [name]."))
+				return
+			if(key.bitt_4 != lock.pin_4)
+				playsound(src, 'mojave/sound/ms13effects/lockpicking/lockpick_tension_11.ogg', 50, TRUE)
+				to_chat(user, span_notice("The key dosen't turn in [name]."))
+				return
+			if(key.bitt_5 != lock.pin_5)
+				playsound(src, 'mojave/sound/ms13effects/lockpicking/lockpick_tension_11.ogg', 50, TRUE)
+				to_chat(user, span_notice("The key dosen't turn in [name]."))
+				return
+			if(key.bitt_6 != lock.pin_6)
+				playsound(src, 'mojave/sound/ms13effects/lockpicking/lockpick_tension_11.ogg', 50, TRUE)
+				to_chat(user, span_notice("The key dosen't turn in [name]."))
+				return
+			if(!lock.item_lock_locked)
+				playsound(src, 'mojave/sound/ms13effects/key_lock.ogg', 50, TRUE)
+				to_chat(user, span_notice("You lock the [name]."))
+				lock.item_lock_locked = TRUE
+				return
+			if(lock.item_lock_locked)
+				playsound(src, 'mojave/sound/ms13effects/key_unlock.ogg', 50, TRUE)
+				to_chat(user, span_notice("You unlock the [name]."))
+				lock.item_lock_locked = FALSE
+				return
 			return
 	//lock interactions
-	if(I.item_flags & LOCKING_ITEM && ms13_flags_1 & LOCKABLE_1)
+	if(I.ms13_flags_1 & LOCKING_ITEM && ms13_flags_1 & LOCKABLE_1)
 		var/obj/item/ms13/lock/potential_lock = I
-		if(lock_locked)
+		if(lock_locked || built_in_lock)
 			to_chat(user, span_warning("The [name] already has a lock."))
 			return
 		if(potential_lock.item_lock_locked)
@@ -129,7 +166,7 @@
 ///
 //for taking the lock off of objects if its open
 /obj/attack_hand_secondary(mob/user, list/modifiers)
-	if(lock)
+	if(lock && !built_in_lock)
 		if(!lock.item_lock_locked && do_after(user, 0.6 SECONDS, src))
 			var/datum/component/lockpickable/lockpickable = GetComponent(/datum/component/lockpickable)
 			qdel(lockpickable)
